refactor: logic, types and utils splitted
